// frontend/src/generated/DraftOperationResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DraftOperationStatus } from "./DraftOperationStatus";
import type { DraftSnapshot } from "./DraftSnapshot";

export type DraftOperationResponse = { status: DraftOperationStatus, code: string, operation_id: string, draft: DraftSnapshot | null, revision: string | null, };